+++ src/internal/connector/exchange/api/api_test.go
@@ -1,0 +1,189 @@
+package api
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"github.com/stretchr/testify/suite"
+
+	"github.com/alcionai/corso/src/internal/connector/graph"
+	"github.com/alcionai/corso/src/internal/tester"
+	"github.com/alcionai/corso/src/pkg/account"
+)
+
+type ExchangeServiceSuite struct {
+	suite.Suite
+	gs          graph.Servicer
+	credentials account.M365Config
+}
+
+func TestExchangeServiceSuite(t *testing.T) {
+	tester.RunOnAny(
+		t,
+		tester.CorsoCITests,
+		tester.CorsoGraphConnectorTests,
+		tester.CorsoGraphConnectorExchangeTests)
+
+	suite.Run(t, new(ExchangeServiceSuite))
+}
+
+func (suite *ExchangeServiceSuite) SetupSuite() {
+	t := suite.T()
+	tester.MustGetEnvSets(t, tester.M365AcctCredEnvs)
+
+	a := tester.NewM365Account(t)
+	m365, err := a.M365Config()
+	require.NoError(t, err)
+
+	suite.credentials = m365
+
+	adpt, err := graph.CreateAdapter(
+		m365.AzureTenantID,
+		m365.AzureClientID,
+		m365.AzureClientSecret)
+	require.NoError(t, err)
+
+	suite.gs = graph.NewService(adpt)
+}
+
+func (suite *ExchangeServiceSuite) TestOptionsForCalendars() {
+	tests := []struct {
+		name       string
+		params     []string
+		checkError assert.ErrorAssertionFunc
+	}{
+		{
+			name:       "Empty Literal",
+			params:     []string{},
+			checkError: assert.NoError,
+		},
+		{
+			name:       "Invalid Parameter",
+			params:     []string{"status"},
+			checkError: assert.Error,
+		},
+		{
+			name:       "Invalid Parameters",
+			params:     []string{"status", "height", "month"},
+			checkError: assert.Error,
+		},
+		{
+			name:       "Valid Parameters",
+			params:     []string{"changeKey", "events", "owner"},
+			checkError: assert.NoError,
+		},
+	}
+	for _, test := range tests {
+		suite.T().Run(test.name, func(t *testing.T) {
+			_, err := optionsForCalendars(test.params)
+			test.checkError(t, err)
+		})
+	}
+}
+
+// TestOptionsForFolders ensures that approved query options
+// are added to the RequestBuildConfiguration. Expected will always be +1
+// on than the input as "id" are always included within the select parameters
+func (suite *ExchangeServiceSuite) TestOptionsForFolders() {
+	tests := []struct {
+		name       string
+		params     []string
+		checkError assert.ErrorAssertionFunc
+		expected   int
+	}{
+		{
+			name:       "Valid Folder Option",
+			params:     []string{"parentFolderId"},
+			checkError: assert.NoError,
+			expected:   2,
+		},
+		{
+			name:       "Multiple Folder Options: Valid",
+			params:     []string{"displayName", "isHidden"},
+			checkError: assert.NoError,
+			expected:   3,
+		},
+		{
+			name:       "Invalid Folder option param",
+			params:     []string{"status"},
+			checkError: assert.Error,
+		},
+	}
+	for _, test := range tests {
+		suite.T().Run(test.name, func(t *testing.T) {
+			config, err := optionsForMailFolders(test.params)
+			test.checkError(t, err)
+			if err == nil {
+				suite.Equal(test.expected, len(config.QueryParameters.Select))
+			}
+		})
+	}
+}
+
+// TestOptionsForContacts similar to TestExchangeService_optionsForFolders
+func (suite *ExchangeServiceSuite) TestOptionsForContacts() {
+	tests := []struct {
+		name       string
+		params     []string
+		checkError assert.ErrorAssertionFunc
+		expected   int
+	}{
+		{
+			name:       "Valid Contact Option",
+			params:     []string{"displayName"},
+			checkError: assert.NoError,
+			expected:   2,
+		},
+		{
+			name:       "Multiple Contact Options: Valid",
+			params:     []string{"displayName", "parentFolderId"},
+			checkError: assert.NoError,
+			expected:   3,
+		},
+		{
+			name:       "Invalid Contact Option param",
+			params:     []string{"status"},
+			checkError: assert.Error,
+		},
+	}
+	for _, test := range tests {
+		suite.T().Run(test.name, func(t *testing.T) {
+			options, err := optionsForContacts(test.params)
+			test.checkError(t, err)
+			if err == nil {
+				suite.Equal(test.expected, len(options.QueryParameters.Select))
+			}
+		})
+	}
+}
+
+// TestGraphQueryFunctions verifies if Query functions APIs
+// through Microsoft Graph are functional
+func (suite *ExchangeServiceSuite) TestGraphQueryFunctions() {
+	ctx, flush := tester.NewContext()
+	defer flush()
+
+	userID := tester.M365UserID(suite.T())
+	tests := []struct {
+		name     string
+		function GraphQuery
+	}{
+		{
+			name:     "GraphQuery: Get All ContactFolders",
+			function: GetAllContactFolderNamesForUser,
+		},
+		{
+			name:     "GraphQuery: Get All Calendars for User",
+			function: GetAllCalendarNamesForUser,
+		},
+	}
+
+	for _, test := range tests {
+		suite.T().Run(test.name, func(t *testing.T) {
+			response, err := test.function(ctx, suite.gs, userID)
+			assert.NoError(t, err)
+			assert.NotNil(t, response)
+		})
+	}
+}
