+++ src/internal/connector/exchange/container_resolver_test.go
@@ -11,8 +11,13 @@ import (
 
 	"github.com/alcionai/corso/src/internal/connector/graph"
 	"github.com/alcionai/corso/src/internal/tester"
+	"github.com/alcionai/corso/src/pkg/account"
 	"github.com/alcionai/corso/src/pkg/path"
 )
+
+// ---------------------------------------------------------------------------
+// mocks and helpers
+// ---------------------------------------------------------------------------
 
 type mockContainer struct {
 	id       *string
@@ -33,6 +38,10 @@ func (m mockContainer) GetDisplayName() *string {
 func (m mockContainer) GetParentFolderId() *string {
 	return m.parentID
 }
+
+// ---------------------------------------------------------------------------
+// unit suite
+// ---------------------------------------------------------------------------
 
 type FolderCacheUnitSuite struct {
 	suite.Suite
@@ -284,6 +293,10 @@ func resolverWithContainers(numContainers int) (*containerResolver, []*mockCache
 	return resolver, containers
 }
 
+// ---------------------------------------------------------------------------
+// configured unit suite
+// ---------------------------------------------------------------------------
+
 // TestConfiguredFolderCacheUnitSuite cannot run its tests in parallel.
 type ConfiguredFolderCacheUnitSuite struct {
 	suite.Suite
@@ -430,4 +443,183 @@ func (suite *ConfiguredFolderCacheUnitSuite) TestAddToCache() {
 	p, err := suite.fc.IDToPath(ctx, m.id)
 	require.NoError(t, err)
 	assert.Equal(t, m.expectedPath, p.String())
+}
+
+// ---------------------------------------------------------------------------
+// integration suite
+// ---------------------------------------------------------------------------
+
+type FolderCacheIntegrationSuite struct {
+	suite.Suite
+	credentials account.M365Config
+	gs          graph.Servicer
+}
+
+func TestFolderCacheIntegrationSuite(t *testing.T) {
+	tester.RunOnAny(
+		t,
+		tester.CorsoCITests,
+		tester.CorsoConnectorExchangeFolderCacheTests)
+
+	suite.Run(t, new(FolderCacheIntegrationSuite))
+}
+
+func (suite *FolderCacheIntegrationSuite) SetupSuite() {
+	t := suite.T()
+	tester.MustGetEnvSets(t, tester.M365AcctCredEnvs)
+
+	a := tester.NewM365Account(t)
+	m365, err := a.M365Config()
+	require.NoError(t, err)
+
+	suite.credentials = m365
+
+	adpt, err := graph.CreateAdapter(
+		m365.AzureTenantID,
+		m365.AzureClientID,
+		m365.AzureClientSecret)
+	require.NoError(t, err)
+
+	suite.gs = graph.NewService(adpt)
+
+	require.NoError(suite.T(), err)
+}
+
+// Testing to ensure that cache system works for in multiple different environments
+func (suite *FolderCacheIntegrationSuite) TestCreateContainerDestination() {
+	ctx, flush := tester.NewContext()
+	defer flush()
+
+	a := tester.NewM365Account(suite.T())
+	m365, err := a.M365Config()
+	require.NoError(suite.T(), err)
+
+	connector, err := createService(m365)
+	require.NoError(suite.T(), err)
+
+	var (
+		user            = tester.M365UserID(suite.T())
+		directoryCaches = make(map[path.CategoryType]graph.ContainerResolver)
+		folderName      = tester.DefaultTestRestoreDestination().ContainerName
+		tests           = []struct {
+			name      string
+			pathFunc1 func(t *testing.T) path.Path
+			pathFunc2 func(t *testing.T) path.Path
+			category  path.CategoryType
+		}{
+			{
+				name:     "Mail Cache Test",
+				category: path.EmailCategory,
+				pathFunc1: func(t *testing.T) path.Path {
+					pth, err := path.Builder{}.Append("Griffindor").
+						Append("Croix").ToDataLayerExchangePathForCategory(
+						suite.credentials.AzureTenantID,
+						user,
+						path.EmailCategory,
+						false,
+					)
+
+					require.NoError(t, err)
+					return pth
+				},
+				pathFunc2: func(t *testing.T) path.Path {
+					pth, err := path.Builder{}.Append("Griffindor").
+						Append("Felicius").ToDataLayerExchangePathForCategory(
+						suite.credentials.AzureTenantID,
+						user,
+						path.EmailCategory,
+						false,
+					)
+
+					require.NoError(t, err)
+					return pth
+				},
+			},
+			{
+				name:     "Contact Cache Test",
+				category: path.ContactsCategory,
+				pathFunc1: func(t *testing.T) path.Path {
+					aPath, err := path.Builder{}.Append("HufflePuff").
+						ToDataLayerExchangePathForCategory(
+							suite.credentials.AzureTenantID,
+							user,
+							path.ContactsCategory,
+							false,
+						)
+
+					require.NoError(t, err)
+					return aPath
+				},
+				pathFunc2: func(t *testing.T) path.Path {
+					aPath, err := path.Builder{}.Append("Ravenclaw").
+						ToDataLayerExchangePathForCategory(
+							suite.credentials.AzureTenantID,
+							user,
+							path.ContactsCategory,
+							false,
+						)
+
+					require.NoError(t, err)
+					return aPath
+				},
+			},
+			{
+				name:     "Event Cache Test",
+				category: path.EventsCategory,
+				pathFunc1: func(t *testing.T) path.Path {
+					aPath, err := path.Builder{}.Append("Durmstrang").
+						ToDataLayerExchangePathForCategory(
+							suite.credentials.AzureTenantID,
+							user,
+							path.EventsCategory,
+							false,
+						)
+					require.NoError(t, err)
+					return aPath
+				},
+				pathFunc2: func(t *testing.T) path.Path {
+					aPath, err := path.Builder{}.Append("Beauxbatons").
+						ToDataLayerExchangePathForCategory(
+							suite.credentials.AzureTenantID,
+							user,
+							path.EventsCategory,
+							false,
+						)
+					require.NoError(t, err)
+					return aPath
+				},
+			},
+		}
+	)
+
+	for _, test := range tests {
+		suite.T().Run(test.name, func(t *testing.T) {
+			folderID, err := CreateContainerDestinaion(
+				ctx,
+				connector,
+				test.pathFunc1(t),
+				folderName,
+				directoryCaches,
+			)
+
+			require.NoError(t, err)
+			resolver := directoryCaches[test.category]
+			_, err = resolver.IDToPath(ctx, folderID)
+			assert.NoError(t, err)
+
+			secondID, err := CreateContainerDestinaion(
+				ctx,
+				connector,
+				test.pathFunc2(t),
+				folderName,
+				directoryCaches,
+			)
+
+			require.NoError(t, err)
+			_, err = resolver.IDToPath(ctx, secondID)
+			require.NoError(t, err)
+			_, ok := resolver.PathInCache(folderName)
+			require.True(t, ok)
+		})
+	}
 }
