+++ src/internal/connector/exchange/service_functions.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/microsoftgraph/msgraph-sdk-go/models"
 	"github.com/pkg/errors"
 
 	"github.com/alcionai/corso/src/internal/connector/graph"
@@ -15,9 +14,6 @@ import (
 
 var ErrFolderNotFound = errors.New("folder not found")
 
-// createService internal constructor for exchangeService struct returns an error
-// iff the params for the entry are incorrect (e.g. len(TenantID) == 0, etc.)
-// NOTE: Incorrect account information will result in errors on subsequent queries.
 func createService(credentials account.M365Config) (*graph.Service, error) {
 	adapter, err := graph.CreateAdapter(
 		credentials.AzureTenantID,
@@ -31,76 +27,7 @@ func createService(credentials account.M365Config) (*graph.Service, error) {
 	return graph.NewService(adapter), nil
 }
 
-// CreateMailFolder makes a mail folder iff a folder of the same name does not exist
-// Reference: https://docs.microsoft.com/en-us/graph/api/user-post-mailfolders?view=graph-rest-1.0&tabs=http
-func CreateMailFolder(ctx context.Context, gs graph.Servicer, user, folder string) (models.MailFolderable, error) {
-	isHidden := false
-	requestBody := models.NewMailFolder()
-	requestBody.SetDisplayName(&folder)
-	requestBody.SetIsHidden(&isHidden)
-
-	return gs.Client().UsersById(user).MailFolders().Post(ctx, requestBody, nil)
-}
-
-func CreateMailFolderWithParent(
-	ctx context.Context,
-	gs graph.Servicer,
-	user, folder, parentID string,
-) (models.MailFolderable, error) {
-	isHidden := false
-	requestBody := models.NewMailFolder()
-	requestBody.SetDisplayName(&folder)
-	requestBody.SetIsHidden(&isHidden)
-
-	return gs.Client().
-		UsersById(user).
-		MailFoldersById(parentID).
-		ChildFolders().
-		Post(ctx, requestBody, nil)
-}
-
-// DeleteMailFolder removes a mail folder with the corresponding M365 ID  from the user's M365 Exchange account
-// Reference: https://docs.microsoft.com/en-us/graph/api/mailfolder-delete?view=graph-rest-1.0&tabs=http
-func DeleteMailFolder(ctx context.Context, gs graph.Servicer, user, folderID string) error {
-	return gs.Client().UsersById(user).MailFoldersById(folderID).Delete(ctx, nil)
-}
-
-// CreateCalendar makes an event Calendar with the name in the user's M365 exchange account
-// Reference: https://docs.microsoft.com/en-us/graph/api/user-post-calendars?view=graph-rest-1.0&tabs=go
-func CreateCalendar(ctx context.Context, gs graph.Servicer, user, calendarName string) (models.Calendarable, error) {
-	requestbody := models.NewCalendar()
-	requestbody.SetName(&calendarName)
-
-	return gs.Client().UsersById(user).Calendars().Post(ctx, requestbody, nil)
-}
-
-// DeleteCalendar removes calendar from user's M365 account
-// Reference: https://docs.microsoft.com/en-us/graph/api/calendar-delete?view=graph-rest-1.0&tabs=go
-func DeleteCalendar(ctx context.Context, gs graph.Servicer, user, calendarID string) error {
-	return gs.Client().UsersById(user).CalendarsById(calendarID).Delete(ctx, nil)
-}
-
-// CreateContactFolder makes a contact folder with the displayName of folderName.
-// If successful, returns the created folder object.
-func CreateContactFolder(
-	ctx context.Context,
-	gs graph.Servicer,
-	user, folderName string,
-) (models.ContactFolderable, error) {
-	requestBody := models.NewContactFolder()
-	temp := folderName
-	requestBody.SetDisplayName(&temp)
-
-	return gs.Client().UsersById(user).ContactFolders().Post(ctx, requestBody, nil)
-}
-
-// DeleteContactFolder deletes the ContactFolder associated with the M365 ID if permissions are valid.
-// Errors returned if the function call was not successful.
-func DeleteContactFolder(ctx context.Context, gs graph.Servicer, user, folderID string) error {
-	return gs.Client().UsersById(user).ContactFoldersById(folderID).Delete(ctx, nil)
-}
-
-// PopulateExchangeContainerResolver gets a folder resolver if one is available for
+// populateExchangeContainerResolver gets a folder resolver if one is available for
 // this category of data. If one is not available, returns nil so that other
 // logic in the caller can complete as long as they check if the resolver is not
 // nil. If an error occurs populating the resolver, returns an error.
